requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
SQLAlchemy==2.0.36
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
